move file, fix update binding
